pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "mail_spreader"
version = "0.1.0"
description = "Automate LinkedIn-based mailings"
authors = [
  { name = "Florian Pasco"}
]
dependencies = [
  "jinja2",
  "pyyaml",
  "typer[all]"
]

[project.scripts]
grader-tools = "mail_spreader.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mail_spreader = ["templates/*.typ", "templates/*.html", "templates/img/*.png"]